+++ lib/src/weather/lib_weather_utils.dart
@@ -46,3 +46,5 @@ List<Interval> makeHistoricalTerm(int startMonth, int endMonth, {int n = 30}) {
 
   return out.sublist(out.length - n);
 }
+
+